% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gp_bonus}
\alias{gp_bonus}
\title{Compute New Grade Points with Bonus}
\usage{
gp_bonus(gp, max.bonus = 15, min.bonus = 0, disallow.A = TRUE)
}
\arguments{
\item{gp}{a vector; student grade points}

\item{max.bonus}{a positive numeric; upper bound of the bonus point}

\item{min.bonus}{a non-negative numeric; lower bound of the bonus point}

\item{disallow.A}{a logical; If \code{TRUE} (default), grade point "A" won't be allowed.}
}
\value{
list; a bonus point and new grade points
}
\description{
Computes student grade points with a bonus by minimizing absolute difference between the fixed significant level 0.05 and a p-value of Kolmogorov-Smirnov test which compares a distribution of grade points with the fixed normal distribution.
}
\details{
It leaves untouched F points which are lower than 60 after the bonus was added.
}
\examples{
## Example 1
gp <- c(43, 72, 88, 60, 77, 51, 69, 61, 60, 86, 65)
gpbonus::gp_bonus(gp)

## Example 2
gp <- c(72, "E", 51, 69, "WF", 81, 61, 75, 54, "W")
gpbonus::gp_bonus(gp)

}
\seealso{
\code{\link{gp_summary}}
}
