% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpbonus-package.r
\docType{package}
\name{gpbonus}
\alias{gpbonus}
\title{Grade Point Bonus}
\description{
Computes student grade points with a bonus by minimizing the absolute difference between the fixed significant level 0.05 and a p-value of Kolmogorov-Smirnov test compares a distribution of grade points with the fixed normal distribution.
}
\author{
MAKHGAL Ganbold
}
